requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
requests>=2.25
numpy>=1.21
vicinity>=0.1.0
sentence-transformers>=2.2.0
torch>=1.13
transformers>=4.20
polars>=0.19
pyarrow>=10.0
pypdf>=3.0
python-docx>=1.0
pypandoc>=1.11
openpyxl>=3.0
ebooklib>=0.18
beautifulsoup4>=4.9
chonkie>=0.6.0 
model2vec>=0.3.0
umap-learn>=0.5
scikit-learn>=1.0
tqdm>=4.60
# tokenlearn>=0.2.0 # for training model2vec